Modulate create page and create a test create
