+++ src/app/create/page.js
@@ -2,56 +2,28 @@
 
 import { useState } from "react";
 import { ChevronLeft } from "lucide-react";
-import Editor from "@monaco-editor/react";
+import BasicInfoForm from "@/components/createlayout/BasicInfoForm";
+import TestCasesForm from "@/components/createlayout/TestCasesForm";
+import RuntimeForm from "@/components/createlayout/RuntimeForm";
 
 /**
- * CreateQuestion component for adding a new coding question.
- *
- * This component allows users to input the question details such as title, difficulty, tags, and coding solution.
- * It also provides a submit button to submit the data to an API endpoint.
- *
- * @component
- * @example
- * return (
- *   <CreateQuestion />
- * );
+ * Helper function to map language names to editor identifiers.
  */
+function mapLanguage(lang) {
+  const normalized = lang.trim().toLowerCase();
+  if (normalized === "c++" || normalized === "c") return "cpp";
+  if (normalized === "javascript") return "javascript";
+  if (normalized === "java") return "java";
+  if (normalized === "python" || normalized === "python3") return "python";
+  return normalized;
+}
+
 export default function CreateQuestion() {
-  /**
-   * State for the question description.
-   * @type {string}
-   */
   const [description, setDescription] = useState("");
-
-  /**
-   * State for the title of the question.
-   * @type {string}
-   */
   const [title, setTitle] = useState("");
-
-  /**
-   * State for difficulty sum.
-   * @type {number}
-   */
   const [difficultySum, setDifficultySum] = useState(5);
-
-  /**
-   * State for checking if we are editing the difficulty bar.
-   * @type {bool}
-   */
   const [editMode, setEditMode] = useState(false);
-
-  /**
-   * State for public display setting.
-   * @type {Array<bool>}
-   */
-
   const [displayPublicly, setDisplayPublicly] = useState(true);
-
-  /**
-   * State for tags.
-   * @type {Object}
-   */
   const [tags, setTags] = useState({
     array: false,
     string: false,
@@ -80,75 +52,14 @@ export default function CreateQuestion() {
     hashing: false,
     bit_mask: false
   });
-
-  const changeTags = (e) => {
-    setTags(e.target.value.split(",").map((tag) => tag.trim()));
-  };
-
-  /**
-   * State for programming languages.
-   * @type {Array<string>}
-   */
   const [languages, setLanguages] = useState([]);
-
-  /**
-   * State for curernt programming language.
-   * @type {string}
-   */
-
   const [selectedLanguage, setSelectedLanguage] = useState("");
-
-  /**
-   * State for code template.
-   * @type {string}
-   */
   const [codeTemplate, setCodeTemplate] = useState("");
-
-  /**
-   * State for code solution.
-   * @type {string}
-   */
   const [codeSolution, setCodeSolution] = useState("");
-
-  /**
-   * State for test cases.
-   * @type {Array<{ANSWER: string, key: string}>}
-   */
-  const [testCases, setTestCases] = useState([{ ANSWER: "", key: "" }]);
-
-  /**
-   * State for active tab.
-   * @type {string}
-   */
+  const [testCases, setTestCases] = useState([{ ANSWER: "", key: {} }]);
   const [activeTab, setActiveTab] = useState("basic-info");
-
-  /**
-   * State for code langauge used.
-   * @type {string}
-   */
   const [codeLanguage, setCodeLanguage] = useState("C++");
 
-  function mapLanguage(lang) {
-    const normalized = lang.trim().toLowerCase();
-    if (normalized === "c++" || normalized === "c") {
-      return "cpp";
-    }
-    if (normalized === "javascript") {
-      return "javascript";
-    }
-    if (normalized === "java") {
-      return "java";
-    }
-    if (normalized === "python" || normalized === "python3") {
-      return "python";
-    }
-    // Return the normalized language if no mapping is needed
-    return normalized;
-  }
-
-  /**
-   * Handles form submission.
-   */
   const handleSubmit = async (e) => {
     e.preventDefault();
 
@@ -172,26 +83,23 @@ export default function CreateQuestion() {
       test_cases: testCases
     };
 
-    // try {
-    //   const response = await fetch("/api/questions/create", {
-    //     method: "POST",
-    //     headers: {
-    //       "Content-Type": "application/json",
-    //     },
-    //     body: JSON.stringify(newQuestion),
-    //   });
+    try {
+      const response = await fetch("/api/questions/create", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newQuestion)
+      });
 
-    //   if (response.ok) {
-    //     alert("Question created successfully!");
-    //   } else {
-    //     const data = await response.json();
-    //     alert(`Error: ${data.error}`);
-    //   }
-    // } catch (error) {
-    //   console.error("Error submitting question:", error);
-    //   alert("Something went wrong, please try again later.");
-    // }
-    console.log(newQuestion);
+      if (response.ok) {
+        alert("Question created successfully!");
+      } else {
+        const data = await response.json();
+        alert(`Error: ${data.error}`);
+      }
+    } catch (error) {
+      console.error("Error submitting question:", error);
+      alert("Something went wrong, please try again later.");
+    }
   };
 
   return (
@@ -203,6 +111,7 @@ export default function CreateQuestion() {
 
       <div className="flex mb-6 space-x-2">
         <button
+          type="button"
           className={`rounded-full px-4 py-2 ${
             activeTab === "basic-info"
               ? "bg-[#4a4a4a]"
@@ -213,6 +122,7 @@ export default function CreateQuestion() {
           Basic Info
         </button>
         <button
+          type="button"
           className={`rounded-full px-4 py-2 ${
             activeTab === "test-cases"
               ? "bg-[#4a4a4a]"
@@ -223,6 +133,7 @@ export default function CreateQuestion() {
           Test Cases
         </button>
         <button
+          type="button"
           className={`rounded-full px-4 py-2 ${
             activeTab === "runtime"
               ? "bg-[#4a4a4a]"
@@ -236,352 +147,50 @@ export default function CreateQuestion() {
 
       <form onSubmit={handleSubmit} className="space-y-6">
         {activeTab === "basic-info" && (
-          <div className="space-y-6">
-            <div className="space-y-2">
-              <label htmlFor="title" className="block text-sm">
-                Title
-              </label>
-              <input
-                id="title"
-                type="text"
-                value={title}
-                onChange={(e) => setTitle(e.target.value)}
-                placeholder="Enter question title"
-                className="w-full bg-[#333333] border-none text-white p-3 rounded"
-              />
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="description" className="block text-sm">
-                Description (markdown supported)
-              </label>
-              <textarea
-                id="description"
-                value={description}
-                onChange={(e) => setDescription(e.target.value)}
-                placeholder="Enter question description"
-                className="w-full min-h-[180px] bg-[#333333] border-none text-white p-3 rounded resize-y"
-              />
-            </div>
-
-            <div className="grid grid-cols-3 gap-4">
-              <div className="flex flex-col items-center">
-                <input
-                  type="range"
-                  min="0.1"
-                  max="10"
-                  step="0.1"
-                  value={difficultySum}
-                  onChange={(e) => setDifficultySum(parseFloat(e.target.value))}
-                  className="w-full"
-                />
-                {editMode ? (
-                  <input
-                    type="number"
-                    step="0.1"
-                    min="0.1"
-                    max="10"
-                    value={difficultySum}
-                    onChange={(e) => {
-                      const val = parseFloat(e.target.value);
-                      setDifficultySum(
-                        isNaN(val)
-                          ? difficultySum
-                          : Math.min(10, Math.max(0.1, val))
-                      );
-                    }}
-                    onBlur={(e) => {
-                      const val = parseFloat(e.target.value);
-                      setDifficultySum(
-                        isNaN(val)
-                          ? difficultySum
-                          : Math.min(10, Math.max(0.1, val))
-                      );
-                      setEditMode(false);
-                    }}
-                    autoFocus
-                    className="mt-2 bg-[#333333] rounded px-3 py-1 w-16 text-center"
-                  />
-                ) : (
-                  <div
-                    className="mt-2 bg-[#333333] rounded px-3 py-1 w-16 text-center cursor-pointer"
-                    onClick={() => setEditMode(true)}
-                  >
-                    {difficultySum.toFixed(1)}
-                  </div>
-                )}
-              </div>
-
-              <div className="space-y-2">
-                <label className="block text-xs text-center">
-                  Display Publicly?
-                  <br />
-                  If unchecked, your question can only be accessed via link
-                </label>
-                <div className="flex justify-center items-center h-12">
-                  <div className="h-6 w-6 rounded bg-[#333333] flex items-center justify-center">
-                    <input
-                      type="checkbox"
-                      checked={displayPublicly}
-                      onChange={(e) => setDisplayPublicly(e.target.checked)}
-                      className="h-6 w-6 rounded bg-[#333333] accent-blue-500"
-                    />
-                  </div>
-                </div>
-              </div>
-
-              <div className="space-y-2">
-                <label className="block text-xs text-center">
-                  How many test cases should
-                  <br />
-                  users see?
-                </label>
-                <div className="flex justify-center">
-                  <input
-                    type="number"
-                    min="0"
-                    defaultValue="3"
-                    className="w-16 text-center bg-[#333333] border-none text-white p-2 rounded"
-                  />
-                </div>
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="languages" className="block text-sm">
-                Programming Languages
-              </label>
-              <div className="flex items-center space-x-2">
-                <select
-                  id="languages"
-                  value={selectedLanguage}
-                  onChange={(e) => setSelectedLanguage(e.target.value)}
-                  className="bg-[#333333] border-none text-white p-3 rounded"
-                >
-                  <option value="">Select a Language</option>
-                  <option value="C++">C++</option>
-                  <option value="Java">Java</option>
-                  <option value="Python">Python</option>
-                  <option value="Python3">Python3</option>
-                  <option value="C">C</option>
-                  <option value="JavaScript">JavaScript</option>
-                </select>
-                <button
-                  type="button"
-                  onClick={() => {
-                    if (
-                      selectedLanguage &&
-                      !languages.includes(selectedLanguage)
-                    ) {
-                      setLanguages([...languages, selectedLanguage]);
-                      setSelectedLanguage("");
-                    }
-                  }}
-                  className="bg-[#4a4a4a] hover:bg-[#5a5a5a] text-white py-2 px-4 rounded"
-                >
-                  Add
-                </button>
-              </div>
-              {languages.length > 0 && (
-                <ul className="mt-2">
-                  {languages.map((lang, idx) => (
-                    <li key={idx} className="flex items-center space-x-2">
-                      <span>{lang}</span>
-                      <button
-                        type="button"
-                        onClick={() =>
-                          setLanguages(languages.filter((l) => l !== lang))
-                        }
-                        className="text-red-500"
-                      >
-                        Remove
-                      </button>
-                    </li>
-                  ))}
-                </ul>
-              )}
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="tags" className="block text-sm">
-                Question Tags
-              </label>
-              <div className="flex flex-wrap gap-2">
-                {Object.keys(tags).map((tag) => (
-                  <button
-                    key={tag}
-                    type="button"
-                    onClick={() =>
-                      setTags({
-                        ...tags,
-                        [tag]: !tags[tag]
-                      })
-                    }
-                    className={`px-3 py-1 rounded ${
-                      tags[tag]
-                        ? "bg-blue-500 text-white"
-                        : "bg-[#333333] text-white"
-                    }`}
-                  >
-                    {tag}
-                  </button>
-                ))}
-              </div>
-            </div>
-          </div>
+          <BasicInfoForm
+            title={title}
+            setTitle={setTitle}
+            description={description}
+            setDescription={setDescription}
+            difficultySum={difficultySum}
+            setDifficultySum={setDifficultySum}
+            editMode={editMode}
+            setEditMode={setEditMode}
+            displayPublicly={displayPublicly}
+            setDisplayPublicly={setDisplayPublicly}
+            languages={languages}
+            setLanguages={setLanguages}
+            selectedLanguage={selectedLanguage}
+            setSelectedLanguage={setSelectedLanguage}
+            tags={tags}
+            setTags={setTags}
+            codeLanguage={codeLanguage}
+            setCodeLanguage={setCodeLanguage}
+          />
         )}
 
         {activeTab === "test-cases" && (
-          <div className="space-y-6">
-            <div className="space-y-2">
-              <label htmlFor="testCases" className="block text-sm">
-                Test Cases (JSON format)
-              </label>
-              <textarea
-                id="testCases"
-                value={JSON.stringify(testCases, null, 2)}
-                onChange={(e) => {
-                  try {
-                    setTestCases(JSON.parse(e.target.value));
-                  } catch (error) {
-                    // Handle JSON parse error
-                  }
-                }}
-                placeholder="Enter test cases in JSON format"
-                className="w-full min-h-[120px] bg-[#333333] border-none text-white p-3 rounded resize-y font-mono text-sm"
-              />
-            </div>
-          </div>
+          <TestCasesForm testCases={testCases} setTestCases={setTestCases} />
         )}
 
         {activeTab === "runtime" && (
-          <div className="space-y-6">
-            <div className="space-y-2">
-              <label className="block text-sm">
-                What Language does this problem use? (you can add additional
-                submission languages after publishing)
-              </label>
-              <select
-                className="w-full bg-[#333333] border-none text-white p-3 rounded"
-                defaultValue="C++"
-                value={codeLanguage}
-                onChange={(e) => setCodeLanguage(e.target.value)}
-              >
-                <option value="C++">C++</option>
-                <option value="Java">Java</option>
-                <option value="Python">Python</option>
-                <option value="Python3">Python3</option>
-                <option value="C">C</option>
-                <option value="JavaScript">JavaScript</option>
-              </select>
-            </div>
-
-            <div className="space-y-2">
-              <label className="block text-sm">
-                Write a function that can parse your Test Cases
-              </label>
-              <div className="border border-gray-700 rounded-lg overflow-hidden shadow-lg">
-                <Editor
-                  id="parseCode"
-                  language={mapLanguage(codeLanguage) || "javascript"}
-                  theme="vs-dark"
-                  defaultValue={`std::vector<void*> parsed;
-void parser() {
-// parsing
-}`}
-                  options={{
-                    fontSize: 14,
-                    minimap: { enabled: false },
-                    automaticLayout: true
-                  }}
-                  className="w-full h-64" // fixed height for better consistency
-                />
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <label className="block text-sm">
-                Write a template function for users to put their solution in
-              </label>
-              <div className="border border-gray-700 rounded-lg overflow-hidden shadow-lg">
-                <Editor
-                  id="codeTemplate"
-                  language={mapLanguage(codeLanguage) || "javascript"}
-                  onChange={(value, event) => setCodeTemplate(value)}
-                  theme="vs-dark"
-                  defaultValue={`void solution() {
-  // code here
-}`}
-                  options={{
-                    fontSize: 14,
-                    minimap: { enabled: false },
-                    automaticLayout: true
-                  }}
-                  className="w-full h-64" // fixed height for better consistency
-                />
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <label className="block text-sm">
-                Write a solution to your problem
-              </label>
-              <div className="border border-gray-700 rounded-lg overflow-hidden shadow-lg">
-                <Editor
-                  id="codeSolution"
-                  language={mapLanguage(codeLanguage) || "javascript"}
-                  onChange={(value, event) => setCodeSolution(value)}
-                  theme="vs-dark"
-                  defaultValue={`void solution() {
-  // code here
-}`}
-                  options={{
-                    fontSize: 14,
-                    minimap: { enabled: false },
-                    automaticLayout: true
-                  }}
-                  className="w-full h-64" // fixed height for better consistency
-                />
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <label className="block text-sm">
-                Finally, write a function that tests the inputs you generated
-                with the user's template function vs your solution. If it passes
-                all test cases, this should only output "all to stdout. If it
-                fails the first test, it should print 1, if it fails the second
-                test, etc.
-              </label>
-
-              <div className="border border-gray-700 rounded-lg overflow-hidden shadow-lg">
-                <Editor
-                  id="parseCode"
-                  language={mapLanguage(codeLanguage) || "javascript"}
-                  theme="vs-dark"
-                  defaultValue={`void testing() {
-  // code here
-}`}
-                  options={{
-                    fontSize: 14,
-                    minimap: { enabled: false },
-                    automaticLayout: true
-                  }}
-                  className="w-full h-64" // fixed height for better consistency
-                />
-              </div>
-            </div>
-          </div>
+          <RuntimeForm
+            codeLanguage={codeLanguage}
+            setCodeLanguage={setCodeLanguage}
+            mapLanguage={mapLanguage}
+            codeTemplate={codeTemplate}
+            setCodeTemplate={setCodeTemplate}
+            codeSolution={codeSolution}
+            setCodeSolution={setCodeSolution}
+          />
         )}
 
-        <div>
-          <button
-            type="submit"
-            className="w-full bg-[#4a4a4a] hover:bg-[#5a5a5a] text-white py-3 px-4 rounded transition-colors"
-          >
-            Create Question
-          </button>
-        </div>
+        <button
+          type="submit"
+          className="w-full bg-[#4a4a4a] hover:bg-[#5a5a5a] text-white py-3 px-4 rounded transition-colors"
+        >
+          Create Question
+        </button>
       </form>
     </div>
   );
